Move BinaryTupleReader
